feat: add performance executer
Added a construct that can be used to generate cold start invocations
for lambda functions.
It is done by running 100 lambda invocations in parrallel through a
express function workflow.
Between the invocation a update to the function is done so that the
next invocation is a cold start as well.
The chunk size, rounds of onvcations and the wait between the chunks
can be configured.